if else nest

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -13,17 +13,17 @@ int main(void)
                     /* your code goes there */
                 if (n > 0)
                 {
-                    printf("%d is positive",n);
+                    printf("%d is positive",(signed int)n);
                 }
                 else
                 {
                     if (n < 0)
                     {
-                        printf("%d is negative",n);
+                        printf("%d is negative",(signed int)n);
                     }
                     else
                     {
-                        printf("%d is zero",n);
+                        printf("%d is zero",(signed int)n);
                     }
                 }
                 return (0);
